extract posture helpers
