+++ web/src/pages/PostureChecksPage/PostureCheckDrawer/PostureCheckDrawer.tsx
@@ -4,7 +4,6 @@ import { useNavigate } from '@tanstack/react-router';
 import { Fragment, Suspense } from 'react';
 import { m } from '../../../paraglide/messages';
 import api from '../../../shared/api/api';
-import type { ApiDevicePosture } from '../../../shared/api/types';
 import { Button } from '../../../shared/defguard-ui/components/Button/Button';
 import { ButtonMenu } from '../../../shared/defguard-ui/components/ButtonMenu/MenuButton';
 import { Chip } from '../../../shared/defguard-ui/components/Chip/Chip';
@@ -19,154 +18,8 @@ import {
 } from '../../../shared/query';
 import { buildPostureCheckMenuItems } from '../postureCheckMenu';
 import type { PostureCheckRow } from '../postureChecks';
+import { buildOsSections } from '../../../shared/utils/postureInfo';
 
-type OsDetailRow = {
-  label: string;
-  value: string | string[];
-};
-
-type OsSection = {
-  icon: (typeof IconKind)[keyof typeof IconKind];
-  name: string;
-  rows: OsDetailRow[];
-};
-
-const getWindowsSection = (
-  rule: Extract<ApiDevicePosture['os_rules'][number], { os_type: 'windows' }>,
-): OsSection => {
-  const otherItems: string[] = [];
-  if (rule.ad_domain_joined_required) otherItems.push('Connected to Active Directory');
-  if (rule.antivirus_required) otherItems.push('Antivirus installed');
-  if (rule.disk_encryption_required) otherItems.push('Disk encryption enabled');
-
-  const rows: OsDetailRow[] = [];
-  if (rule.min_os_version !== null) {
-    rows.push({ label: 'Version', value: `Windows ${rule.min_os_version} and higher` });
-  }
-  if (rule.windows_security_update_max_age !== null) {
-    rows.push({
-      label: 'Update',
-      value: `Minimum ${rule.windows_security_update_max_age} month update`,
-    });
-  }
-  if (otherItems.length > 0) {
-    rows.push({ label: 'Other', value: otherItems });
-  }
-
-  return { icon: IconKind.Windows, name: 'Windows', rows };
-};
-
-const getMacosSection = (
-  rule: Extract<ApiDevicePosture['os_rules'][number], { os_type: 'macos' }>,
-): OsSection => {
-  const otherItems: string[] = [];
-  if (rule.disk_encryption_required) otherItems.push('Disk encryption enabled');
-  if (rule.device_integrity_required) otherItems.push('Device integrity');
-
-  const rows: OsDetailRow[] = [];
-  if (rule.min_os_version !== null) {
-    rows.push({ label: 'Version', value: `macOS ${rule.min_os_version} and higher` });
-  }
-  if (otherItems.length > 0) {
-    rows.push({ label: 'Other', value: otherItems });
-  }
-
-  return { icon: IconKind.Apple, name: 'macOS', rows };
-};
-
-const getLinuxSection = (
-  rule: Extract<ApiDevicePosture['os_rules'][number], { os_type: 'linux' }>,
-): OsSection => {
-  const otherItems: string[] = [];
-  if (rule.disk_encryption_required) otherItems.push('Disk encryption enabled');
-
-  const rows: OsDetailRow[] = [];
-  if (rule.min_kernel_version !== null) {
-    rows.push({
-      label: 'Version',
-      value: `Kernel ${rule.min_kernel_version} and higher`,
-    });
-  }
-  if (otherItems.length > 0) {
-    rows.push({ label: 'Other', value: otherItems });
-  }
-
-  return { icon: IconKind.Linux, name: 'Linux', rows };
-};
-
-const getIosSection = (
-  rule: Extract<ApiDevicePosture['os_rules'][number], { os_type: 'ios' }>,
-): OsSection => {
-  const rows: OsDetailRow[] = [];
-  if (rule.min_os_version !== null) {
-    rows.push({ label: 'Version', value: `iOS ${rule.min_os_version}+` });
-  }
-
-  return { icon: IconKind.AppStore, name: 'iOS', rows };
-};
-
-const getAndroidSection = (
-  rule: Extract<ApiDevicePosture['os_rules'][number], { os_type: 'android' }>,
-): OsSection => {
-  const otherItems: string[] = [];
-  if (rule.device_integrity_required) otherItems.push('Device integrity');
-
-  const rows: OsDetailRow[] = [];
-  if (rule.min_os_version !== null) {
-    rows.push({ label: 'Version', value: `Android ${rule.min_os_version}+` });
-  }
-  if (otherItems.length > 0) {
-    rows.push({ label: 'Other', value: otherItems });
-  }
-
-  return { icon: IconKind.Android, name: 'Android', rows };
-};
-
-const getDefguardSection = (posture: ApiDevicePosture): OsSection | null => {
-  if (!posture.min_client_version && !posture.allow_prerelease_client) return null;
-
-  const rows: OsDetailRow[] = [];
-  if (posture.min_client_version) {
-    rows.push({
-      label: 'Version',
-      value: `Defguard ${posture.min_client_version} and higher`,
-    });
-  }
-  if (posture.allow_prerelease_client) {
-    rows.push({ label: 'Other', value: 'Pre-release allowed' });
-  }
-
-  return { icon: IconKind.Defguard, name: 'Defguard', rows };
-};
-
-const buildOsSections = (posture: ApiDevicePosture): OsSection[] => {
-  const sections: OsSection[] = [];
-
-  for (const rule of posture.os_rules ?? []) {
-    switch (rule.os_type) {
-      case 'windows':
-        sections.push(getWindowsSection(rule));
-        break;
-      case 'macos':
-        sections.push(getMacosSection(rule));
-        break;
-      case 'linux':
-        sections.push(getLinuxSection(rule));
-        break;
-      case 'ios':
-        sections.push(getIosSection(rule));
-        break;
-      case 'android':
-        sections.push(getAndroidSection(rule));
-        break;
-    }
-  }
-
-  const defguardSection = getDefguardSection(posture);
-  if (defguardSection) sections.push(defguardSection);
-
-  return sections;
-};
 
 type ContentProps = {
   row: PostureCheckRow;
